Revision 6, v2.0
Rolled Back Input Data Stream/Parseing Changes. Will Need To Work On Tomorrow.

+++ header/roster.h
@@ -16,7 +16,9 @@ class Roster {
         // TODO: Find out if there is a way to dynamically set this variable. 
             // IDEA: May have to do a static initial variable. Then count the array items in the input. Assign that count total to the int numStudents. Then perform a new array creation and delete the initial one or see if there is a way to count the input before the class call then pass the details. 
         
-        int numStudents = 0;
+        // Changing from int to __int64 to remove compiler warning regarding possible data loss due to type conversion. 
+
+        __int64 numStudents = 0;
 
         // Changing from int to __int64 to remove compiler warning regarding possible data loss due to type conversion. 
         __int64 inputSize = 0;
@@ -65,6 +67,9 @@ class Roster {
         // TODO: DELETE FROM FINAL -- CREATED FOR TESTING
         Student* getStudent(int i);
 
+
         // MEMBER FUNCTIONS - END
         // ==== ==== ==== ==== ==== ==== ==== ==== 
+
+
 };
